+++ releasenotes.md
@@ -9,8 +9,8 @@
   * The mouse pointer may now be used to move the text caret while entering a savegame description in the save game menu.
   * Navigating menus using the mouse pointer is now fixed when the `vid_widescreen` CVAR is `off`.
 * The positions of `ON`/`OFF` and `HIGH`/`LOW` in the options menu have been fixed in some instances.
-* The player’s health, armor and ammo in the status bar and widescreen HUDs now animate when they change. This can be disabled by the new `animatedstats` CVAR, which is `on` by default and `off` when vanilla mode is enabled.
 * These changes have been made to the alternate widescreen HUD:
+  * The player’s health, armor and ammo now animate when they change.
   * The position of the player’s health is now fixed when 0% and the `negativehealth` CVAR is `off`.
   * The vertical positions of some of the weapon silhouettes have changed.
 * Crashes no longer occur when: